Working version (to be tested)
24/09/2020
This is the first working version of the package. The very basic functions are implemented. You can compute the error vectors, the cosine maps, run a permutation analysis, and estimate the parameters of the CAM model.
There are a couple of tutorials with a recangle and a circle shapes (see the Tutorials folder) . In future, I will need to test the toolbox with other types of shapes or objects.