Working version
I have fixed the psychtoolbox rendering problem (I have re-installed  new versions of gstreamer and psychtoolbox)

There is working demo now (Demo\run_prototypes_experiment_Circle.m). There are still many adjustments to do, but this is a good starting point. One of the next things to add is the triggers for the GSV.