Implemented haptic condition
The haptic condition is now embedded in prototypes_show_oneTrial.m. This can be used by setting .Modality='haptic' and useCamera=1
This script needs further testing, but it seems ok.
Another relevant change was the addition of the column "dot_id" which was done directly in test_prototypes_generate_grid.m. This needs to be checked as well.