Relevant updates
prototypes_generate_trials_order_new:
major changes, so keep an eye on this. It is now simpler to generate ABBA blocks. This requires new parameter (nrep, blocks_order) . Also, the grid is not repeated when generated, but only here

set_camera:
need to be tested

split_StringPairs:
just an utility function