Major changes!
I have updated the names of the variables in a Prototypes table. The easiest way to see which names need to be used is looking in the file func\data\prototypes_check_prototable.m.

Some of the existing data and the experiment script might need to be updated. I have created a function called func\data\prototypes_update_dataFormat.m that can help with that. This is called in a script called func\Tutorials\FIXDATASET.m. This scripts will be moved in future updates.
Other updates have been done in the plotting functions, but only to ensure they work now. In future, these might need an update as well.

The tutorial tutorial01_CosineMaps_Rectangle.m and related data are working now. In future updates, I should also test the other tutorials and data. The functions prototypes_fit_model and prototypes_model_CAM seem to work fine.